created script selector
